refactor: make porfoliocards for value and funds in component and reusable extending simplewidget
