--- src/main/java/edu/ntnu/idatt2003/millions/view/dashboard/portfolio/card/AvailableFundsCard.java
@@ -1,27 +1,0 @@
-package edu.ntnu.idatt2003.millions.view.dashboard.portfolio.card;
-
-import edu.ntnu.idatt2003.millions.service.GameService;
-import edu.ntnu.idatt2003.millions.util.CurrencyFormatter;
-import edu.ntnu.idatt2003.millions.view.component.StyledText;
-import edu.ntnu.idatt2003.millions.view.component.WidgetCard;
-
-/**
- * Widget card displaying the player's currently available funds for trading.
- */
-public class AvailableFundsCard extends WidgetCard {
-
-    private final GameService gameService;
-    private final StyledText valueLabel = StyledText.widgetValue();
-
-    public AvailableFundsCard(GameService gameService) {
-        super(gameService, "dashboard.availableFunds");
-        this.gameService = gameService;
-        getChildren().addAll(titleLabel, valueLabel);
-        refreshDisplay();
-    }
-
-    @Override
-    protected void refreshDisplay() {
-        valueLabel.setText(CurrencyFormatter.format(gameService.getPlayer().getMoney()));
-    }
-}
